added class info, armor info, and weapon info
Hey everybody, I got all the information for the armor. class, and
weapons in the book now. I also re-centered the text that was located on
the bottom of the front page. The class information starts on page 4 and
ends on page 38. the weapon information starts on page 40 and ends on
page 44. the armor information starts on page 45 and ends on page 49.